+++ src/createStore.ts
@@ -12,16 +12,23 @@ const validateStateKey = (keys: string[], stateKey: string) => {
 };
 
 type Store<State, Action> = {
+  useStoreState: <StateKey extends keyof State>(stateKey: StateKey) => State[StateKey];
+  /**
+   * @deprecated useStoreState instead
+   */
   useGlobalState: <StateKey extends keyof State>(stateKey: StateKey) => readonly [State[StateKey]];
   getState: () => State;
   dispatch: (action: Action) => Action;
 };
 
 /**
- * create a global store
+ * Create a global store.
  *
- * In additon to `useGlobaState` which is the same hook as in createGlobalState,
- * a store has `getState` and `dispatch`.
+ * It returns a set of functions
+ * - `useStoreState`: a custom hook to read store state by key
+ * - `getState`: a function to get store state by key outside React
+ * - `dispatch`: a function to dispatch an action to store
+ *
  * A store works somewhat similarly to Redux, but not the same.
  *
  * @example
@@ -31,10 +38,10 @@ type Store<State, Action> = {
  * const reducer = ...;
  *
  * const store = createStore(reducer, initialState);
- * const { useGlobalState } = store;
+ * const { useStoreState, dispatch } = store;
  *
  * const Component = () => {
- *   const [count, setCount] = useGlobalState('count');
+ *   const count = useStoreState('count');
  *   ...
  * };
  */
@@ -50,16 +57,24 @@ export const createStore = <State extends object, Action extends { type: unknown
   type StateKeys = keyof State;
   const keys = Object.keys(initialState);
 
-  const useGlobalState = <StateKey extends StateKeys>(stateKey: StateKey) => {
+  const useStoreState = <StateKey extends StateKeys>(stateKey: StateKey) => {
     if (process.env.NODE_ENV !== 'production') {
       validateStateKey(keys, stateKey as string);
     }
     const selector = useCallback((state: State) => state[stateKey], [stateKey]);
-    const partialState = useStore(selector);
-    return [partialState] as const;
+    return useStore(selector);
+  };
+
+  const useGlobalState = <StateKey extends StateKeys>(stateKey: StateKey) => {
+    if (process.env.NODE_ENV !== 'production') {
+      // eslint-disable-next-line no-console
+      console.warn('[DEPRECATED] useStoreState instead');
+    }
+    return [useStoreState(stateKey)] as const;
   };
 
   return {
+    useStoreState,
     useGlobalState,
     getState: useStore.getState,
     dispatch: (useStore as any).dispatch,